Implemented NSFetchResultsController
Functionality is implemented in a little bit of a hacky way because the
NSFetchedResultsController was not properly updating after the first
download of wod entries.  It's working here, but just because I manually
create a new NSFetchedResultsController and re-fetch the results anytime
wods are downloaded.  I really should be able to just rely on the
NSFetechedResultsControllerDelegate methods to do this.